Initial test: Add Menu Font Scaling
This feature is not ready for use but I have added a temporary slider to the Dev Settings tab.  The value is a scale multiplier for all of the font values used when building the font.  This cannot be updated during runtime as it will result in a crash.

Some UI elements need to be multiplied by this scalar as well.