Revert "Merge branch 'dev/adunn/fix_decals' into 'main'"
This reverts commit 16ce0dffaf8254255c26cafecbe91146ab38b623, reversing
changes made to c02d00eab4f51a611aacf50ee89395f440df0d61.

This change is being reverted temporarily because the merge commit belonging to this MR suffered from a bug in our default message template system.  As such, it renders the message useless and we'd like to correct that.