GPU Perf: Payload compaction and quality improvement
Perf impact is pretty small for this one now. Helps 2% in some cases but not much in others after merging other optimizations. Only measured on 4090.

However, in compacting the payloads, we enable some nice other quality benefits around rayprecision.  We can represent a more directions on the sphere. This is particularly noticable in primary rays, where previously saw tooths are sometimes visible as geometry aliasing.