[REMIX-3574] Make sure we disable POM in situations where its not used
The resolve function today correctly handles cases where a path vertex resolve terminates on an opaque hit, but there are other cases to consider - like when the hit is not opaque, or there is no hit. In such cases we should ensure we return a false to pomOpaqueSurfaceEncountered.

Also eliminates unnecessary branched loads for surface index when executing POM checks in direct/indirect/nee lighting paths.