Fix for incorrectly ignoring meshes after show/hide is used in USD Composer
Previously, a mesh that is overridden but has no child meshes will be ignored, regardless of the state of the references or the presence of the preserveOriginalDrawCall attribute.

This changes that behavior so that override prims with no children that do not explicitly override the references will preserve the original draw call.