[REMIX-3919] fix shimmering in translucent surfaces when using Transformer RR
Combining the Attenuation with the Albedo before passing to RR was resulting in nearly-black pixels, which causes errors in RR.

We instead extract a color tint from the attenuation by normalizing it, avoiding including any dimming effect.  This prevents the RR bug.
To reduce the chances of the bug reoccurring, we also clamp near-black pixels to just be black.