[REMIX-4105] simplifying and standardizing the RtxOption API
Changes the RTX_OPTION macro to declare the RtxOption struct using `optionName`, adds a () operator for that struct, and a `.set()` function.

Removes the `optionNameDescriptor()` function, since it can be easily accessed as `optionName.getDescriptor()`, and was rarely used.

Removes the ability to access the value by using `m_optionName`, as the implicit type cast can cause confusion (and is also rarely used).