System Information Report Logging
* Add a system information report logging function to log information about the CPU, memory and OS that Remix is running on to provide slightly more information in log files to make debugging easier.
* Improve byte format string helper.

This is mainly a response to GitHub tickets which often require asking users which CPU Remix is being run on (e.g. to determine if integrated graphics are potentially causing issues) due to that information not being included in the log files. There's still more information which could be enumerated here (e.g. all Nvidia/AMD/Intel devices + driver versions on the system rather than just the device Vulkan selects), but this should be a good start for now to address the main lacking bit of information in bug reports.