[REMIX-4112] Implement Rtx Option Clamping
adds min and max values for int, float, and vector types. These are applied for all types of sets (code, imgui, .conf file, and API sets), and can be changed at runtime.

This migrates all of the uses of RTX_OPTION_CLAMP that happen after the initial setup, since those were all broken. Uses in the first frame are not migrated, but should be in the future.